install libcurl4-openssl-dev
libcurl's development headers (whatever that is) are required by the r package curl.

The actual error message is this:
Package libcurl was not found in the pkg-config search path.
Perhaps you should add the directory containing `libcurl.pc'
to the PKG_CONFIG_PATH environment variable

Apparently, libcurl *is* installed in ubuntu-latest but its development headers are not. See here: https://git.io/JR2se

That is why I assumed that the r package curl required these headers. I might be wrong, of course.